Move JacksonRequestRecordBuffer to http_json.jackson package
